avoid compiler warning (implicit cast from double to float)
SVN-Revision: 29583

+++ contrib/brl/bseg/boxm/algo/boxm_compute_scene_difference.h
@@ -16,7 +16,6 @@
 #include <vcl_iostream.h>
 
 
-
 void boxm_compute_block_difference(boxm_block<boct_tree<short, boxm_sample<BOXM_APM_MOG_GREY> > > * block1,
                                    boxm_block<boct_tree<short, boxm_sample<BOXM_APM_MOG_GREY> > > * block2,
                                    boxm_block<boct_tree<short, boxm_sample<BOXM_APM_MOG_GREY> > > * blockout,
@@ -25,7 +24,7 @@ void boxm_compute_block_difference(boxm_block<boct_tree<short, boxm_sample<BOXM_
   typedef boxm_sample<BOXM_APM_MOG_GREY> data_type;
   typedef boct_tree<short, boxm_sample<BOXM_APM_MOG_GREY> > tree_type;
   typedef boct_tree_cell<short, boxm_sample<BOXM_APM_MOG_GREY> > cell_type;
-  
+
   tree_type* tree1 = block1->get_tree();
   tree_type* tree2 = block2->get_tree();
   tree_type* treeout = blockout->get_tree();
@@ -33,37 +32,37 @@ void boxm_compute_block_difference(boxm_block<boct_tree<short, boxm_sample<BOXM_
   vcl_vector<cell_type*> leaves=tree1->leaf_cells();
   vcl_vector<cell_type*> leavesout=treeout->leaf_cells();
 
-  for(unsigned i=0;i<leaves.size();i++)
+  for (unsigned i=0;i<leaves.size();i++)
   {
     cell_type *cell1=leaves[i];
     cell_type *cellout=leavesout[i];
     vgl_point_3d<double> p=cell1->local_bounding_box(tree1->root_level()).centroid();
     cell_type *cell2=tree2->locate_point(p);
-    if(cell2)
+    if (cell2)
     {
-        data_type data1=cell1->data();
-        float len1=tree1->cell_bounding_box(cell1).width();
+      data_type data1=cell1->data();
+      double len1=tree1->cell_bounding_box(cell1).width();
 
-        data_type data2=cell2->data();
-        float len2=tree2->cell_bounding_box(cell2).width();
+      data_type data2=cell2->data();
+      double len2=tree2->cell_bounding_box(cell2).width();
 
-        data_type dataout=cellout->data();
+      data_type dataout=cellout->data();
 
-        float p1=1-vcl_exp(-data1.alpha*len1);
-        float p2=1-vcl_exp(-data2.alpha*len2);
+      double p1=1-vcl_exp(-data1.alpha*len1);
+      double p2=1-vcl_exp(-data2.alpha*len2);
 
-        float p=p1+p2-2*p1*p2;
+      double p=p1+p2-2*p1*p2;
 
-        float alphaout=-vcl_log(1-p)/len1;
-        if(p<threshold)
-            alphaout=0.0f;
-        dataout.alpha=alphaout;
-        if(p1>p2)
-            dataout.appearance_=data1.appearance_;
-        else
-            dataout.appearance_=data2.appearance_;
+      double alphaout=-vcl_log(1-p)/len1;
+      if (p<threshold)
+          alphaout=0.0;
+      dataout.alpha=(float)alphaout;
+      if (p1>p2)
+        dataout.appearance_=data1.appearance_;
+      else
+        dataout.appearance_=data2.appearance_;
 
-        cellout->set_data(dataout);
+      cellout->set_data(dataout);
     }
   }
 }
@@ -83,17 +82,17 @@ void boxm_compute_scene_difference(boxm_scene<boct_tree<short, boxm_sample<BOXM_
   boxm_block_iterator<tree_type> iterout(&sceneout);
 
   for (; !iter1.end(); iter1++,iter2++,iterout++) {
-      vcl_cout<<".";
-      scene1.load_block(iter1.index());
-      scene2.load_block(iter2.index());
-      sceneout.load_block(iterout.index());
+    vcl_cout<<'.';
+    scene1.load_block(iter1.index());
+    scene2.load_block(iter2.index());
+    sceneout.load_block(iterout.index());
 
-      //boxm_block<tree_type>* block1 = *iter1;
-      //boxm_block<tree_type>* block2 = *iter2;
-      //boxm_block<tree_type>* blockout = *iterout;
+    //boxm_block<tree_type>* block1 = *iter1;
+    //boxm_block<tree_type>* block2 = *iter2;
+    //boxm_block<tree_type>* blockout = *iterout;
 
-      boxm_compute_block_difference((*iter1),(*iter2),(*iterout), threshold);
-      sceneout.write_active_block();
+    boxm_compute_block_difference((*iter1),(*iter2),(*iterout), threshold);
+    sceneout.write_active_block();
   }
 }
